% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_first_function.R
\name{my_first_function}
\alias{my_first_function}
\title{My first function}
\usage{
my_first_function(inp)
}
\arguments{
\item{inp}{the input, a vector}
}
\value{
A list of the values where x are the inputs and y is the square of the inputs
}
\description{
My first function
}
\examples{
my_first_function(1:10)
}
