% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residualsumsquares.R
\name{RSS}
\alias{RSS}
\title{Residual Sum Squares}
\usage{
RSS(model, yv)
}
\arguments{
\item{model}{a regression model}

\item{yv}{the y values (dependent)}
}
\value{
the residual sum square, aka the sum of square of the residuals
}
\description{
Residual Sum Squares
}
\examples{
Data_Frame <- data.frame (
Pulse = c(100, 150, 120, 120),
Duration = c(60, 30, 45, 40)
)
Data_Frame
model <- lm(Data_Frame$Pulse ~ Data_Frame$Duration)
RSS(model, Data_Frame$Pulse)
}
